[2022-07-18] 업로드 관리 - Content로 변경
